:truck: move Layout to container and pass state down
