+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import Head from "next/head";
 import { ThemeProvider } from "styled-components";
 
 import GlobalStyle from "@/styles/global";
-import Layout from "@/components/Layout";
+import Layout from "@/containers/Layout";
 import theme from "@/styles/theme";
 
 const App: FC<AppProps> = ({ Component, pageProps }) => (
